+++ changelog.md
@@ -21,7 +21,8 @@
 * Redirect factomd `current-minute` API call to v2 path (intead of debug path).
 * Increased performance of `getEntryWithBlockContext`: fixed cost of 3 factomd API calls instead of rewinding the entry blocks of the chain.
 * `keyToPrivateFctAddress` renamed `seedToPrivateFctAddress`.
-* `keyToPrivateEcAddress` renamed `seedToPrivateEcAddress`
+* `keyToPrivateEcAddress` renamed `seedToPrivateEcAddress`.
+* `add` throws an Error when trying to create a chain that already exists.
 
 ### Removed